Move world gen into its own folder
